+++ android/src/main/java/ar/com/intrale/mobile/mercadopago/MercadopagoPlugin.java
@@ -10,6 +10,12 @@ import java.util.logging.Logger;
 
 import com.mercadopago.android.px.core.MercadoPagoCheckout;
 
+import com.mercadopago.android.px.internal.callbacks.CallbackHolder;
+import com.mercadopago.android.px.internal.callbacks.PaymentCallback;
+import com.mercadopago.android.px.internal.callbacks.PaymentDataCallback;
+import com.mercadopago.android.px.model.Payment;
+import com.mercadopago.android.px.model.PaymentData;
+import com.mercadopago.android.px.model.exceptions.MercadoPagoError;
 import io.flutter.embedding.engine.plugins.FlutterPlugin;
 import io.flutter.embedding.engine.plugins.activity.ActivityAware;
 import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding;
@@ -40,63 +46,67 @@ public class MercadopagoPlugin implements FlutterPlugin, MethodCallHandler, Acti
 
   @Override @SuppressWarnings("unchecked")
   public void onAttachedToEngine(@NonNull FlutterPluginBinding flutterPluginBinding) {
-    LOGGER.info("start onAttachedToEngine");
+    LOGGER.info("MercadopagoPlugin: start onAttachedToEngine");
     channel = new MethodChannel(flutterPluginBinding.getBinaryMessenger(), "mercadopago");
     channel.setMethodCallHandler(this);
     context = flutterPluginBinding.getApplicationContext();
-    LOGGER.info("end onAttachedToEngine");
+    LOGGER.info("MercadopagoPlugin: end onAttachedToEngine");
   }
 
   @Override @SuppressWarnings("unchecked")
   public void onMethodCall(@NonNull MethodCall call, @NonNull Result result) {
-    LOGGER.info("start onMethodCall");
+    LOGGER.info("MercadopagoPlugin: start onMethodCall");
     if (call.method.equals("startPayment")) {
-      LOGGER.info("call startPayment");
+      LOGGER.info("MercadopagoPlugin: call startPayment");
       startPayment(call, result);
-      result.success("Android " + android.os.Build.VERSION.RELEASE);
+      //result.success("Android " + android.os.Build.VERSION.RELEASE);
     } else {
       result.notImplemented();
     }
-    LOGGER.info("end onMethodCall");
-  }
-
-  @Override @SuppressWarnings("unchecked")
-  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
-    LOGGER.info("start onDetachedFromEngine");
-    channel.setMethodCallHandler(null);
-    LOGGER.info("end onDetachedFromEngine");
+    LOGGER.info("MercadopagoPlugin: end onMethodCall");
   }
 
   @SuppressWarnings("unchecked")
   private void startPayment(MethodCall call, Result result) {
-    LOGGER.info("start startPayment");
+    LOGGER.info("MercadopagoPlugin: start startPayment");
+
     HashMap<String, Object> arguments = (HashMap<String, Object>) call.arguments;
     String publicKey = (String) arguments.get(MERCADO_PAGO_PUBLIC_KEY_ARG);
     String preferenceId = (String) arguments.get(MERCADO_PAGO_PREFERENCE_ID_ARG);
-    new MercadoPagoCheckout.Builder(publicKey, preferenceId)
-            .build().startPayment(context, MERCADO_PAGO_REQUEST_CODE);
-    LOGGER.info("end startPayment");
+    LOGGER.info("MercadopagoPlugin: Build Mercado pago => publicKey:" + publicKey + ", preferenceId:" + preferenceId);
+
+    IntraleMercadoPagoCheckout checkout = new IntraleMercadoPagoCheckout(publicKey, preferenceId);
+    checkout.startPayment(context, MERCADO_PAGO_REQUEST_CODE, result);
+
+    LOGGER.info("MercadopagoPlugin: end startPayment");
   }
-  
+
+  @Override @SuppressWarnings("unchecked")
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+    LOGGER.info("MercadopagoPlugin: start onDetachedFromEngine");
+    channel.setMethodCallHandler(null);
+    LOGGER.info("MercadopagoPlugin: end onDetachedFromEngine");
+  }
+
   @Override
   public void  onAttachedToActivity(ActivityPluginBinding binding ) {
-    LOGGER.info("start onAttachedToActivity");
+    LOGGER.info("MercadopagoPlugin: start onAttachedToActivity");
     activity = binding.getActivity();
-    LOGGER.info("end onAttachedToActivity");
+    LOGGER.info("MercadopagoPlugin: end onAttachedToActivity");
 	}
   
   @Override
   public void  onDetachedFromActivity() {
-      //TODO("Not yet implemented")
+    LOGGER.info("MercadopagoPlugin: onDetachedFromActivity");
   }
   
   @Override
   public void onReattachedToActivityForConfigChanges(ActivityPluginBinding binding) {
-      //TODO("Not yet implemented")
+    LOGGER.info("MercadopagoPlugin: onReattachedToActivityForConfigChanges");
   }
   
   @Override
   public void  onDetachedFromActivityForConfigChanges() {
-	    //TODO("Not yet implemented")
+    LOGGER.info("MercadopagoPlugin: onDetachedFromActivityForConfigChanges");
   }
 }